move some code after extracting functionality out of the account controller
