+++ DDDEastAnglia.Tests/Admin/UserControllerTests.cs
@@ -2,6 +2,7 @@
 using DDDEastAnglia.Areas.Admin.Controllers;
 using DDDEastAnglia.DataAccess;
 using DDDEastAnglia.Helpers;
+using DDDEastAnglia.Helpers.LoginMethods;
 using DDDEastAnglia.Models;
 using NSubstitute;
 using NUnit.Framework;
