+++ DDDEastAnglia/Areas/Admin/Controllers/UserController.cs
@@ -4,6 +4,7 @@ using System.Web.Mvc;
 using DDDEastAnglia.Areas.Admin.Models;
 using DDDEastAnglia.DataAccess;
 using DDDEastAnglia.Helpers;
+using DDDEastAnglia.Helpers.LoginMethods;
 using DDDEastAnglia.Models;
 
 namespace DDDEastAnglia.Areas.Admin.Controllers
