+++ DDDEastAnglia/DDDEastAnglia.csproj
@@ -227,6 +227,10 @@
     <Compile Include="Helpers\LoginMethodViewModelBuilder.cs" />
     <Compile Include="Helpers\Sessions\SelectedSessions.cs" />
     <Compile Include="Helpers\OauthLoginIconProvider.cs" />
+    <Compile Include="Helpers\LoginMethods\ExternalLoginsProvider.cs" />
+    <Compile Include="Helpers\LoginMethods\LoginMethodLoader.cs" />
+    <Compile Include="Helpers\LoginMethods\LoginMethodViewModelBuilder.cs" />
+    <Compile Include="Helpers\LoginMethods\LoginMethodIconProvider.cs" />
     <Compile Include="Models\LoginMethod.cs" />
     <Compile Include="DataAccess\SimpleData\Queries\AccountLoginMethodQuery.cs" />
     <Compile Include="DataAccess\SpeakerRepository.cs" />
