+++ DDDEastAnglia/Helpers/LoginMethods/LoginMethodLoader.cs
@@ -5,7 +5,7 @@ using DDDEastAnglia.DataAccess;
 using DDDEastAnglia.DataAccess.Builders;
 using DDDEastAnglia.Models;
 
-namespace DDDEastAnglia.Helpers
+namespace DDDEastAnglia.Helpers.LoginMethods
 {
     public interface ILoginMethodLoader
     {
